Move asyncStates to top-level helper, use in ProjectHome component
