+++ packages/app-project/components/ProjectHome/ProjectHome.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { inject, observer } from 'mobx-react'
+import asyncStates from '../../helpers/asyncStates'
 
 function storeMapper (stores) {
   return {
@@ -11,7 +12,7 @@ function storeMapper (stores) {
 @observer
 class ProjectHome extends React.Component {
   render () {
-    if (this.props.project.state === 'loading') {
+    if (this.props.project.loadingState === asyncStates.loading) {
       return (
         <div>Loading</div>
       )
